fix a few warnings

+++ src/lib/core/wavelet/WaveletReverse.cpp
@@ -1489,8 +1489,9 @@ void WaveletReverse::v_16_53(const dwt_scratch<int16_t>* scratch, Buffer2dSimple
         for(uint32_t c = 0; c < nb_cols; c++, winL.buf_++, winH.buf_++, winDest.buf_++)
         {
           scratch->mem[1] = (int16_t)(winL.buf_[0] - ((winH.buf_[0] + 1) >> 1));
-          winDest.buf_[0] = std::clamp<int16_t>(winH.buf_[0] + scratch->mem[1] + dc, dcMin, dcMax);
-          winDest.buf_[1] = std::clamp<int16_t>(scratch->mem[1] + dc, dcMin, dcMax);
+          winDest.buf_[0] =
+              (int16_t)std::clamp<int32_t>(winH.buf_[0] + scratch->mem[1] + dc, dcMin, dcMax);
+          winDest.buf_[1] = (int16_t)std::clamp<int32_t>(scratch->mem[1] + dc, dcMin, dcMax);
         }
       }
       else
